requirements.txt
django
flask-cors
setuptools
crispy-bootstrap4
pillow
django-crispy-forms
djangorestframework